# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BiznisBox API",
        "description": "BiznisBox API Documentation",
        "contact": {
            "email": "info@biznisbox.com"
        },
        "license": {
            "name": "AGPL-3.0-or-later",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "2.0.0-beta"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "otp": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@test.com",
                                    "otp": "123456",
                                    "password": "useruser1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "response_time": {
                                            "type": "string",
                                            "example": "1.2ms"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJzdWIiOjEsImlzcyI6Imh0dHA6Ly9sb2NhbGhvc3Q6ODAwMC9hcGkvdXNlci9sb2dpbiIsImlhdCI6MTYxNzQwNjYwMiwiZXhwIjoxNjE3NDA2NjQyLCJuYmYiOjE2MTc0MDY2NDIsImp0aSI6Ij"
                                                },
                                                "expires_in": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-04-06 14:00:42"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "response_time": {
                                            "type": "string",
                                            "example": "1.2ms"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "response_time": {
                                            "type": "string",
                                            "example": "1.2ms"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "dd"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Me",
                "description": "Me",
                "operationId": "Me",
                "responses": {
                    "200": {
                        "description": "Successfully data retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "response_time": {
                                            "type": "string",
                                            "example": "1.2ms"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status_code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "9cbbf943-32b4-412f-89ee-e82e18d01e0f"
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "picture": {
                                                    "type": "string",
                                                    "example": "9cbbf943-32b4-412f-89ee-e82e18d01e0f.jpg"
                                                },
                                                "last_login_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-04-06 14:00:42"
                                                },
                                                "language": {
                                                    "type": "string",
                                                    "example": "en"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "UTC"
                                                },
                                                "theme": {
                                                    "type": "string",
                                                    "example": "light"
                                                },
                                                "two_factor_auth": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "avatar_url": {
                                                    "type": "string",
                                                    "example": "http://example.com/avatar.jpg"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "super_admin"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "super_admin"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}